Add cluster and node memory metrics

+++ pkg/proxmox/types.go
@@ -50,5 +50,5 @@ type RootFs struct {
 type Swap struct {
 	Free  int64 `json:"free"`
 	Total int64 `json:"total"`
-	Used  int   `json:"used"`
+	Used  int64 `json:"used"`
 }
